convert department admin to typescript
