+++ app/javascript/DepartmentAdmin/NewDepartment.tsx
