--- app/javascript/DepartmentAdmin/buildDepartmentInput.js
@@ -1,6 +1,0 @@
-export default function buildDepartmentInput(department) {
-  return {
-    name: department.name,
-    proposal_description: department.proposal_description,
-  };
-}
